Rename `CustomTourProvider` to `AppTourProvider`
