+++ src/components/StudyList/StudyList.tsx
@@ -20,7 +20,7 @@ import NoneOr from "../NoneOr/NoneOr";
 import QueryErrorBanner from "../QueryErrorBanner/QueryErrorBanner";
 import { StepType } from "@reactour/tour";
 import { useLocalTour } from "../CustomTourProvider/hooks";
-import { TourId } from "../CustomTourProvider/CustomTourProvider";
+import { TourId } from "../CustomTourProvider/AppTourProvider";
 
 // Make steps for the tour.
 // Reference: https://developer.mozilla.org/en-US/docs/Learn/HTML/Howto/Use_data_attributes
